requirements.txt
boto3==1.4.4
botocore==1.5.86
certifi==2017.4.17
chardet==3.0.4
coverage==4.2
dj-database-url==0.3.0
Django==1.11
django-extensions==1.6.1
django-test-without-migrations==0.4
docutils==0.13.1
idna==2.5
Jinja2==2.9.6
jmespath==0.9.3
MarkupSafe==1.0
model-mommy==1.3.2
pkg-resources==0.0.0
psycopg2==2.6.1
pycep-correios==2.2.0
python-dateutil==2.6.1
python-decouple==3.0
pytz==2017.2
requests==2.18.1
s3transfer==0.1.10
six==1.10.0
urllib3==1.21.1
python-thumbnails==0.5.1